LPS-134746 Move upgrade to web and populate resource actions at upgrade start.
