+++ modules/apps/change-tracking/change-tracking-service/src/main/java/com/liferay/change/tracking/internal/upgrade/ChangeTrackingServiceUpgrade.java
@@ -17,13 +17,10 @@ package com.liferay.change.tracking.internal.upgrade;
 import com.liferay.change.tracking.internal.upgrade.v2_2_0.CTPreferencesUpgradeProcess;
 import com.liferay.change.tracking.internal.upgrade.v2_3_0.UpgradeCompanyId;
 import com.liferay.change.tracking.internal.upgrade.v2_4_0.CTSchemaVersionUpgradeProcess;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Daniel Kocsis
@@ -60,21 +57,7 @@ public class ChangeTrackingServiceUpgrade implements UpgradeStepRegistrator {
 			new com.liferay.change.tracking.internal.upgrade.v2_5_0.
 				SchemaUpgradeProcess());
 
-		registry.register(
-			"2.5.0", "2.5.1",
-			new com.liferay.change.tracking.internal.upgrade.v2_5_1.
-				PublicationsUserRoleUpgradeProcess(
-					_resourcePermissionLocalService, _roleLocalService,
-					_userLocalService));
+		registry.register("2.5.0", "2.5.1", new DummyUpgradeStep());
 	}
-
-	@Reference
-	private ResourcePermissionLocalService _resourcePermissionLocalService;
-
-	@Reference
-	private RoleLocalService _roleLocalService;
-
-	@Reference
-	private UserLocalService _userLocalService;
 
 }
